Fixing UTF8 bugs in writing to the stream
More UTF8 corrections during the writing process, yet again I was using the wrong length variable which was corrected by converting to UTF8 and then using the length of the resulting bytes. Should now work for UTF8 values, if you must change the encoding (for the non-English web developers, simply perform a search or replace for Encoding.UTF8 and change it to Encoding.UTF16 or Encoding.UTF32.