Security fix: possible infinite loop
It was possible for a malicious client to cause an infinite loop, greatly effecting performance (but not denying service), via not sending a \r\n sequence. This has been fixed by only polling during header reading "READRETRIES" times (a constant at the time of the Server.cs file), at which point it'll return a 444. I originally did not want to do this due to performance issues.